+++ R/scplot-generic.R
@@ -1,0 +1,4 @@
+#' @export
+scplot <- function(object, ...) {
+  UseMethod("scplot")
+}